Bug 8764: Fix handling of old negative
(cherry picked from commit ed78dd32e7dac1d8aa74cb637fcdc549aff4e627)

+++ mappingservice/implementation/src/main/java/org/opendaylight/lispflowmapping/implementation/lisp/MapServer.java
@@ -10,6 +10,7 @@ package org.opendaylight.lispflowmapping.implementation.lisp;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Maps;
+import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import java.net.InetAddress;
 import java.net.NetworkInterface;
@@ -103,9 +104,8 @@ public class MapServer implements IMapServerAsync, OdlMappingserviceListener, IS
     @SuppressWarnings("unchecked")
     public void handleMapRegister(MapRegister mapRegister) {
         boolean mappingUpdated = false;
-        boolean oldMappingRemoved = false;
         boolean merge = ConfigIni.getInstance().mappingMergeIsSet() && mapRegister.isMergeEnabled();
-        Set<Subscriber> subscribers = null;
+        Set<Subscriber> subscribers = Sets.newConcurrentHashSet();
         MappingRecord oldMapping;
 
         if (merge) {
@@ -126,9 +126,21 @@ public class MapServer implements IMapServerAsync, OdlMappingserviceListener, IS
 
             oldMapping = getMappingRecord(mapService.getMapping(MappingOrigin.Southbound, eid));
             mapService.addMapping(MappingOrigin.Southbound, eid, getSiteId(mapRegister), mappingData);
-            if (oldMapping != null && MappingRecordUtil.isNegativeMapping(oldMapping)) {
+            if (oldMapping != null
+                    && MappingRecordUtil.isNegativeMapping(oldMapping)
+                    && !oldMapping.getEid().equals(eid)) {
+                if (subscriptionService) {
+                    // Here we save the subscribers of the OLD mapping before removing. We will add to this set the
+                    // subscribers of the NEW mapping below (since the EIDs are different, the result of
+                    // mappingChanged() will be true, and then send an SMR to all subscribers with the EID of the NEW
+                    // mapping only.
+                    Set<Subscriber> oldMappingSubscribers = getSubscribers(oldMapping.getEid());
+                    if (oldMappingSubscribers != null) {
+                        subscribers.addAll(oldMappingSubscribers);
+                        LoggingUtil.logSubscribers(LOG, oldMapping.getEid(), subscribers);
+                    }
+                }
                 mapService.removeMapping(MappingOrigin.Southbound, oldMapping.getEid());
-                oldMappingRemoved = true;
             }
 
             if (subscriptionService) {
@@ -138,17 +150,17 @@ public class MapServer implements IMapServerAsync, OdlMappingserviceListener, IS
                 if (mappingChanged(oldMapping, newMapping)) {
                     if (LOG.isDebugEnabled()) {
                         LOG.debug("Mapping update occured for {} SMRs will be sent for its subscribers.",
-                                LispAddressStringifier.getString(mapping.getEid()));
+                                LispAddressStringifier.getString(eid));
                     }
-                    subscribers = getSubscribers(eid);
+                    Set<Subscriber> newMappingSubscribers = getSubscribers(eid);
                     if (oldMapping != null && !oldMapping.getEid().equals(eid)) {
-                        subscribers = addParentSubscribers(eid, subscribers);
+                        newMappingSubscribers = addParentSubscribers(eid, newMappingSubscribers);
+                    }
+                    if (newMappingSubscribers != null) {
+                        subscribers.addAll(newMappingSubscribers);
+                        LoggingUtil.logSubscribers(LOG, eid, subscribers);
                     }
                     handleSmr(eid, subscribers);
-                    if (oldMapping != null && oldMappingRemoved && !oldMapping.getEid().equals(eid)) {
-                        subscribers = getSubscribers(oldMapping.getEid());
-                        handleSmr(oldMapping.getEid(), subscribers);
-                    }
                     mappingUpdated = true;
                 }
             }
